+++ frontend/src/app/order-printed-book/[id]/page.tsx
@@ -6,7 +6,7 @@ import useGetBook from "@components/hooks/useGetBook";
 import { useFormik } from "formik";
 import ReusableInput from "../../../components/ReusableInput";
 import { useAuth0 } from "@auth0/auth0-react";
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState, MutableRefObject } from "react";
 import usePrintBook from "@components/hooks/usePrintBook";
 import Image from "next/image";
 import axios from "axios";
@@ -26,308 +26,48 @@ interface ShippingDetailsValues {
 const OrderBook = ({ params }: { params: OrderBookValues }) => {
   const id = Number(params.id);
   const { isLoading, isError, bookData } = useGetBook(id);
-  const { data } = usePrintBook();
-  console.log(bookData);
-
-  const [selectedBookSize, setSelectedBookSize] = useState<string>("");
-
-  const [interiorColors, setInteriorColors] = useState<string[]>([]);
-  const [selectedInteriorColor, setSelectedInteriorColor] =
-    useState<string>("");
-  const [selectedColor, setSelectedColor] = useState<string>("");
-  const [selectedPrintQuality, setSelectedPrintQuality] = useState<string>("");
-
-  const [paperTypes, setPaperTypes] = useState<string[]>([]);
-  const [selectedPaperType, setSelectedPaperType] = useState<string>("");
-
-  const [bindingTypes, setBindingTypes] = useState<string[]>([]);
-  const [selectedBindingType, setSelectedBindingType] = useState<string>("");
-
-  const [coverFinishes, setCoverFinishes] = useState<string[]>([]);
-  const [selectedCoverFinish, setSelectedCoverFinish] = useState<string>("");
-
-  const [sku, setSku] = useState<string | null>(null);
-
-  const [quantity, setQuantity] = useState<number | null>(null);
-
-  const [country, setCountry] = useState<string>("");
-
-  const [streetAddress, setStreetAddress] = useState<string>("");
-  const [city, setCity] = useState<string>("");
-  const [state, setState] = useState<string>("");
-  const [postalCode, setPostalCode] = useState<string>("");
-  const [email, setEmail] = useState<string>("");
-  const [phoneNumber, setPhoneNumber] = useState<string>("");
-  const [countryCode, setCountryCode] = useState<string>("");
-
-  const [costs, setCosts] = useState<any>(null);
-
-  const [selectedShippingOption, setSelectedShippingOption] = useState<
-    string | null
-  >(null);
-  const [shippingOptions, setShippingOptions] = useState<any>([]);
+  const {
+    data,
+    selectedBookSize,
+    selectedInteriorColor,
+    selectedColor,
+    selectedPrintQuality,
+    selectedPaperType,
+    selectedBindingType,
+    selectedCoverFinish,
+    shippingOptions,
+    selectedShippingOption,
+    countryCode,
+    countryList,
+    quantity,
+    fullName,
+    streetAddress,
+    city,
+    state,
+    postalCode,
+    email,
+    phoneNumber,
+    country,
+    costs,
+    podPackageId,
+    handleBookSizeChange,
+    handleInteriorColorChange,
+    handlePaperTypeChange,
+    handleBindingTypeChange,
+    handleCoverFinishChange,
+    getShippingOptions,
+    handleEstimatedCost,
+    handleNumberOfCopiesChange,
+    handleShippingOptionChange,
+  } = usePrintBook(bookData?.page!);
 
   const { user, getAccessTokenSilently } = useAuth0();
 
   const pageCount = bookData?.page!;
 
-  const shippingAddress =
-    streetAddress && city && state && postalCode && email && phoneNumber;
+  const countryNames: string[] = Object.keys(countryList);
 
-  const countryList = {
-    Afghanistan: "AF",
-    Albania: "AL",
-    Algeria: "DZ",
-    "American Samoa": "AS",
-    Andorra: "AD",
-    Angola: "AO",
-    Anguilla: "AI",
-    Antarctica: "AQ",
-    "Antigua and Barbuda": "AG",
-    Argentina: "AR",
-    Armenia: "AM",
-    Aruba: "AW",
-    Australia: "AU",
-    Austria: "AT",
-    Azerbaijan: "AZ",
-    "Bahamas (the)": "BS",
-    Bahrain: "BH",
-    Bangladesh: "BD",
-    Barbados: "BB",
-    Belarus: "BY",
-    Belgium: "BE",
-    Belize: "BZ",
-    Benin: "BJ",
-    Bermuda: "BM",
-    Bhutan: "BT",
-    "Bolivia (Plurinational State of)": "BO",
-    "Bonaire, Sint Eustatius and Saba": "BQ",
-    "Bosnia and Herzegovina": "BA",
-    Botswana: "BW",
-    "Bouvet Island": "BV",
-    Brazil: "BR",
-    "British Indian Ocean Territory (the)": "IO",
-    "Brunei Darussalam": "BN",
-    Bulgaria: "BG",
-    "Burkina Faso": "BF",
-    Burundi: "BI",
-    "Cabo Verde": "CV",
-    Cambodia: "KH",
-    Cameroon: "CM",
-    Canada: "CA",
-    "Cayman Islands (the)": "KY",
-    "Central African Republic (the)": "CF",
-    Chad: "TD",
-    Chile: "CL",
-    China: "CN",
-    "Christmas Island": "CX",
-    "Cocos (Keeling) Islands (the)": "CC",
-    Colombia: "CO",
-    "Comoros (the)": "KM",
-    "Congo (the Democratic Republic of the)": "CD",
-    "Congo (the)": "CG",
-    "Cook Islands (the)": "CK",
-    "Costa Rica": "CR",
-    Croatia: "HR",
-    Cuba: "CU",
-    Curaçao: "CW",
-    Cyprus: "CY",
-    Czechia: "CZ",
-    "Côte d'Ivoire": "CI",
-    Denmark: "DK",
-    Djibouti: "DJ",
-    Dominica: "DM",
-    "Dominican Republic (the)": "DO",
-    Ecuador: "EC",
-    Egypt: "EG",
-    "El Salvador": "SV",
-    "Equatorial Guinea": "GQ",
-    Eritrea: "ER",
-    Estonia: "EE",
-    Eswatini: "SZ",
-    Ethiopia: "ET",
-    "Falkland Islands (the) [Malvinas]": "FK",
-    "Faroe Islands (the)": "FO",
-    Fiji: "FJ",
-    Finland: "FI",
-    France: "FR",
-    "French Guiana": "GF",
-    "French Polynesia": "PF",
-    "French Southern Territories (the)": "TF",
-    Gabon: "GA",
-    "Gambia (the)": "GM",
-    Georgia: "GE",
-    Germany: "DE",
-    Ghana: "GH",
-    Gibraltar: "GI",
-    Greece: "GR",
-    Greenland: "GL",
-    Grenada: "GD",
-    Guadeloupe: "GP",
-    Guam: "GU",
-    Guatemala: "GT",
-    Guernsey: "GG",
-    Guinea: "GN",
-    "Guinea-Bissau": "GW",
-    Guyana: "GY",
-    Haiti: "HT",
-    "Heard Island and McDonald Islands": "HM",
-    "Holy See (the)": "VA",
-    Honduras: "HN",
-    "Hong Kong": "HK",
-    Hungary: "HU",
-    Iceland: "IS",
-    India: "IN",
-    Indonesia: "ID",
-    "Iran (Islamic Republic of)": "IR",
-    Iraq: "IQ",
-    Ireland: "IE",
-    "Isle of Man": "IM",
-    Israel: "IL",
-    Italy: "IT",
-    Jamaica: "JM",
-    Japan: "JP",
-    Jersey: "JE",
-    Jordan: "JO",
-    Kazakhstan: "KZ",
-    Kenya: "KE",
-    Kiribati: "KI",
-    "Korea (the Democratic People's Republic of)": "KP",
-    "Korea (the Republic of)": "KR",
-    Kuwait: "KW",
-    Kyrgyzstan: "KG",
-    "Lao People's Democratic Republic (the)": "LA",
-    Latvia: "LV",
-    Lebanon: "LB",
-    Lesotho: "LS",
-    Liberia: "LR",
-    Libya: "LY",
-    Liechtenstein: "LI",
-    Lithuania: "LT",
-    Luxembourg: "LU",
-    Macao: "MO",
-    Madagascar: "MG",
-    Malawi: "MW",
-    Malaysia: "MY",
-    Maldives: "MV",
-    Mali: "ML",
-    Malta: "MT",
-    "Marshall Islands (the)": "MH",
-    Martinique: "MQ",
-    Mauritania: "MR",
-    Mauritius: "MU",
-    Mayotte: "YT",
-    Mexico: "MX",
-    "Micronesia (Federated States of)": "FM",
-    "Moldova (the Republic of)": "MD",
-    Monaco: "MC",
-    Mongolia: "MN",
-    Montenegro: "ME",
-    Montserrat: "MS",
-    Morocco: "MA",
-    Mozambique: "MZ",
-    Myanmar: "MM",
-    Namibia: "NA",
-    Nauru: "NR",
-    Nepal: "NP",
-    "Netherlands (the)": "NL",
-    "New Caledonia": "NC",
-    "New Zealand": "NZ",
-    Nicaragua: "NI",
-    "Niger (the)": "NE",
-    Nigeria: "NG",
-    Niue: "NU",
-    "Norfolk Island": "NF",
-    "Northern Mariana Islands (the)": "MP",
-    Norway: "NO",
-    Oman: "OM",
-    Pakistan: "PK",
-    Palau: "PW",
-    "Palestine, State of": "PS",
-    Panama: "PA",
-    "Papua New Guinea": "PG",
-    Paraguay: "PY",
-    Peru: "PE",
-    "Philippines (the)": "PH",
-    Pitcairn: "PN",
-    Poland: "PL",
-    Portugal: "PT",
-    "Puerto Rico": "PR",
-    Qatar: "QA",
-    "Republic of North Macedonia": "MK",
-    Romania: "RO",
-    "Russian Federation (the)": "RU",
-    Rwanda: "RW",
-    Réunion: "RE",
-    "Saint Barthélemy": "BL",
-    "Saint Helena, Ascension and Tristan da Cunha": "SH",
-    "Saint Kitts and Nevis": "KN",
-    "Saint Lucia": "LC",
-    "Saint Martin (French part)": "MF",
-    "Saint Pierre and Miquelon": "PM",
-    "Saint Vincent and the Grenadines": "VC",
-    Samoa: "WS",
-    "San Marino": "SM",
-    "Sao Tome and Principe": "ST",
-    "Saudi Arabia": "SA",
-    Senegal: "SN",
-    Serbia: "RS",
-    Seychelles: "SC",
-    "Sierra Leone": "SL",
-    Singapore: "SG",
-    "Sint Maarten (Dutch part)": "SX",
-    Slovakia: "SK",
-    Slovenia: "SI",
-    "Solomon Islands": "SB",
-    Somalia: "SO",
-    "South Africa": "ZA",
-    "South Georgia and the South Sandwich Islands": "GS",
-    "South Sudan": "SS",
-    Spain: "ES",
-    "Sri Lanka": "LK",
-    "Sudan (the)": "SD",
-    Suriname: "SR",
-    "Svalbard and Jan Mayen": "SJ",
-    Sweden: "SE",
-    Switzerland: "CH",
-    "Syrian Arab Republic": "SY",
-    "Taiwan (Province of China)": "TW",
-    Tajikistan: "TJ",
-    "Tanzania, United Republic of": "TZ",
-    Thailand: "TH",
-    "Timor-Leste": "TL",
-    Togo: "TG",
-    Tokelau: "TK",
-    Tonga: "TO",
-    "Trinidad and Tobago": "TT",
-    Tunisia: "TN",
-    Turkey: "TR",
-    Turkmenistan: "TM",
-    "Turks and Caicos Islands (the)": "TC",
-    Tuvalu: "TV",
-    Uganda: "UG",
-    Ukraine: "UA",
-    "United Arab Emirates (the)": "AE",
-    "United Kingdom of Great Britain and Northern Ireland (the)": "GB",
-    "United States Minor Outlying Islands (the)": "UM",
-    "United States of America (the)": "US",
-    Uruguay: "UY",
-    Uzbekistan: "UZ",
-    Vanuatu: "VU",
-    "Venezuela (Bolivarian Republic of)": "VE",
-    "Viet Nam": "VN",
-    "Virgin Islands (British)": "VG",
-    "Virgin Islands (U.S.)": "VI",
-    "Wallis and Futuna": "WF",
-    "Western Sahara": "EH",
-    Yemen: "YE",
-    Zambia: "ZM",
-    Zimbabwe: "ZW",
-    "Åland Islands": "AX",
-  };
-  const countryNames = Object.keys(countryList);
-
-  let bookSizes: string[] = Array.from(
+  const bookSizes: string[] = Array.from(
     new Set(
       data
         .filter((data: Array<string | number>) => {
@@ -339,199 +79,88 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
     )
   );
 
-  const handleBookSizeChange = (size: string) => {
-    setSelectedBookSize(size);
-    setSelectedInteriorColor("");
-    setSelectedColor("");
-    setSelectedPrintQuality("");
-    setSelectedPaperType("");
-    setSelectedBindingType("");
-    setSelectedCoverFinish("");
+  const interiorColors: string[] = Array.from(
+    new Set(
+      data
+        .filter((data: Array<string | number>) => {
+          return (
+            data[2] === selectedBookSize &&
+            +data[3] < pageCount &&
+            +data[4] > pageCount
+          );
+        })
+        .map((data: Array<string | number>) => {
+          return data[23] + "_" + data[24];
+        })
+    )
+  );
 
-    const sizeFilteredData = data.filter((data: Array<string | number>) => {
-      return data[2] === size && +data[3] < pageCount && +data[4] > pageCount;
-    });
+  const paperTypes: string[] = Array.from(
+    new Set(
+      data
+        .filter((data: Array<string | number>) => {
+          return (
+            data[2] === selectedBookSize &&
+            +data[3] < pageCount &&
+            +data[4] > pageCount &&
+            data[23] === selectedColor &&
+            data[24] === selectedPrintQuality
+          );
+        })
+        .map((data: Array<string | number>) => {
+          return data[31];
+        })
+    )
+  );
 
-    setInteriorColors(
-      Array.from(
-        new Set(
-          sizeFilteredData.map((data: Array<string | number>) => {
-            return data[23] + "_" + data[24];
-          })
-        )
-      )
-    );
+  const bindingTypes: string[] = Array.from(
+    new Set(
+      data
+        .filter((data: Array<string | number>) => {
+          return (
+            data[2] === selectedBookSize &&
+            +data[3] < pageCount &&
+            +data[4] > pageCount &&
+            data[23] === selectedColor &&
+            data[24] === selectedPrintQuality &&
+            data[31] === selectedPaperType
+          );
+        })
+        .map((data: Array<string | number>) => {
+          return data[25];
+        })
+    )
+  );
 
-    setPaperTypes([]);
-    setBindingTypes([]);
-    setCoverFinishes([]);
-    setSku(null);
-  };
+  const coverFinishes: string[] = Array.from(
+    new Set(
+      data
+        .filter((data: Array<string | number>) => {
+          return (
+            data[2] === selectedBookSize &&
+            +data[3] < pageCount &&
+            +data[4] > pageCount &&
+            data[23] === selectedColor &&
+            data[24] === selectedPrintQuality &&
+            data[31] === selectedPaperType &&
+            data[25] === selectedBindingType
+          );
+        })
+        .map((data: Array<string | number>) => {
+          return data[32];
+        })
+    )
+  );
 
-  const handleInteriorColorChange = (interiorColor: string) => {
-    setSelectedInteriorColor(interiorColor);
-    setSelectedColor(interiorColor.split("_")[0]);
-    setSelectedPrintQuality(interiorColor.split("_")[1]);
-    setSelectedPaperType("");
-    setSelectedBindingType("");
-    setSelectedCoverFinish("");
-
-    const interiorColorFilterData = data.filter(
-      (data: Array<string | number>) => {
-        return (
-          data[2] === selectedBookSize &&
-          +data[3] < pageCount &&
-          +data[4] > pageCount &&
-          data[23] === interiorColor.split("_")[0] &&
-          data[24] === interiorColor.split("_")[1]
-        );
-      }
-    );
-
-    setPaperTypes(
-      Array.from(
-        new Set(
-          interiorColorFilterData.map((data: Array<string | number>) => {
-            return data[31];
-          })
-        )
-      )
-    );
-    setBindingTypes([]);
-    setCoverFinishes([]);
-    setSku(null);
-  };
-
-  const handlePaperTypeChange = (paperType: string) => {
-    setSelectedPaperType(paperType);
-    setSelectedBindingType("");
-    setSelectedCoverFinish("");
-
-    const paperFilterData = data.filter((data: Array<string | number>) => {
-      return (
-        data[2] === selectedBookSize &&
-        +data[3] < pageCount &&
-        +data[4] > pageCount &&
-        data[23] === selectedColor &&
-        data[24] === selectedPrintQuality &&
-        data[31] === paperType
-      );
-    });
-
-    setBindingTypes(
-      Array.from(
-        new Set(
-          paperFilterData.map((data: Array<string | number>) => {
-            return data[25];
-          })
-        )
-      )
-    );
-    setCoverFinishes([]);
-    setSku(null);
-  };
-
-  const handleBindingTypeChange = (bindingType: string) => {
-    setSelectedBindingType(bindingType);
-    setSelectedCoverFinish("");
-
-    const bindFilterData = data.filter((data: Array<string | number>) => {
-      return (
-        data[2] === selectedBookSize &&
-        +data[3] < pageCount &&
-        +data[4] > pageCount &&
-        data[23] === selectedColor &&
-        data[24] === selectedPrintQuality &&
-        data[31] === selectedPaperType &&
-        data[25] === bindingType
-      );
-    });
-
-    setCoverFinishes(
-      Array.from(
-        new Set(
-          bindFilterData.map((data: Array<string | number>) => {
-            return data[32];
-          })
-        )
-      )
-    );
-    setSku(null);
-  };
-
-  const handleCoverFinishChange = (coverFinish: string) => {
-    setSelectedCoverFinish(coverFinish);
-
-    const coverFinishFilteredData = data.filter(
-      (data: Array<string | number>) => {
-        return (
-          data[2] === selectedBookSize &&
-          +data[3] < pageCount &&
-          +data[4] > pageCount &&
-          data[23] === selectedColor &&
-          data[24] === selectedPrintQuality &&
-          data[31] === selectedPaperType &&
-          data[25] === selectedBindingType &&
-          data[32] === coverFinish
-        );
-      }
-    );
-
-    setSku(coverFinishFilteredData[0][0]);
-  };
-
-  const getShippingOptions = async (country: keyof typeof countryList) => {
-    setCountryCode(countryList[`${country}`]);
-
-    const response = await axios.post(
-      "https://api.lulu.com/shipping-options/",
-      {
-        currency: "USD",
-        line_items: [
-          {
-            page_count: pageCount,
-            pod_package_id: sku,
-            quantity: quantity,
-          },
-        ],
-        shipping_address: {
-          // city: city,
-          country: countryList[`${country}`],
-          // postcode: postalCode,
-          // state_code: state,
-          // street1: streetAddress,
-        },
-      }
-    );
-    setShippingOptions(
-      response.data.filter((f: any) => {
-        return f.is_active === true;
-      })
-    );
-  };
-
-  const handleEstimatedCost = async (shippingOption: string) => {
-    setSelectedShippingOption(shippingOption);
-
-    const response = await axiosInstance.post(
-      "/api/orders/printJobEstimatedCost",
-      {
-        pageCount,
-        podPackageId: sku,
-        quantity,
-        city,
-        countryCode,
-        postalCode,
-        state,
-        streetAddress,
-        shippingLevel: shippingOption,
-        phoneNumber,
-      }
-    );
-    console.log(response.data.costs);
-
-    setCosts(response.data.costs);
-  };
+  const printingCost =
+    Math.round(costs?.line_item_costs[0]?.total_cost_incl_tax * 100) / 100;
+  const shippingCost =
+    Math.round(
+      (+costs?.shipping_cost?.total_cost_incl_tax +
+        +costs?.fulfillment_cost?.total_cost_incl_tax) *
+        100
+    ) / 100;
+  const totalCost = Math.round((printingCost + shippingCost) * 100) / 100;
 
   const makePayment = async () => {
     const token = await getAccessTokenSilently();
@@ -542,24 +171,18 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
         {
           bookID: id,
           userID: user?.sub,
-          totalAmount:
-            Math.round(
-              (+costs?.line_item_costs[0]?.total_cost_incl_tax +
-                +costs?.shipping_cost?.total_cost_incl_tax +
-                +costs?.fulfillment_cost?.total_cost_incl_tax) *
-                100
-            ) / 100,
-          quantity,
           bookData,
-          podPackageId: sku,
-          email,
-          name: "abcd",
-          city,
+          totalAmount: totalCost,
+          quantity,
+          podPackageId: podPackageId[0],
+          email: email.current?.value,
+          name: fullName.current?.value,
+          city: email.current?.value,
           countryCode,
-          phoneNumber,
-          postalCode,
-          stateCode: state,
-          streetAddress,
+          phoneNumber: phoneNumber.current?.value,
+          postalCode: postalCode.current?.value,
+          stateCode: state.current?.value,
+          streetAddress: streetAddress.current?.value,
           shippingLevel: selectedShippingOption,
         },
         {
@@ -1082,116 +705,6 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
         </div>
 
         <div className="mt-10">
-          <div className="flex flex-col gap-5">
-            <h1 className="text-xl lg:text-2xl font-quicksand font-semibold">
-              Shipping Address
-            </h1>
-            <hr className="border-1 border-black"></hr>
-
-            <div className="md:flex flex-col mt-5">
-              <label
-                htmlFor="streetAddress"
-                className="text-xl lg:text-2xl font-quicksand"
-              >
-                Street Address <span className="asterisk">*</span>
-              </label>
-              <input
-                type="text"
-                name="streetAddress"
-                id="streetAddress"
-                value={streetAddress}
-                onChange={(e) => setStreetAddress(e.target.value)}
-                className="border outline-none focus:border-slate-600 py-2 rounded-md"
-              />
-            </div>
-
-            <div className="flex gap-10">
-              <div className="w-1/2 md:flex flex-col">
-                <label
-                  htmlFor="city"
-                  className="text-xl lg:text-2xl font-quicksand"
-                >
-                  City <span className="asterisk">*</span>
-                </label>
-                <input
-                  type="text"
-                  name="city"
-                  id="city"
-                  value={city}
-                  onChange={(e) => setCity(e.target.value)}
-                  className="border outline-none focus:border-slate-600 py-2 rounded-md"
-                />
-              </div>
-              <div className="w-1/2 md:flex flex-col">
-                <label
-                  htmlFor="state"
-                  className="text-xl lg:text-2xl font-quicksand"
-                >
-                  State <span className="asterisk">*</span>
-                </label>
-                <input
-                  type="text"
-                  name="state"
-                  id="state"
-                  value={state}
-                  onChange={(e) => setState(e.target.value)}
-                  className="border outline-none focus:border-slate-600 py-2 rounded-md"
-                />
-              </div>
-            </div>
-
-            <div className="flex justify-between gap-5 my-5">
-              <div className="w-1/3 md:flex flex-col">
-                <label
-                  htmlFor="email"
-                  className="text-xl lg:text-2xl font-quicksand"
-                >
-                  Email <span className="asterisk">*</span>
-                </label>
-                <input
-                  type="email"
-                  name="email"
-                  id="email"
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  className="border outline-none focus:border-slate-600 py-2 rounded-md"
-                />
-              </div>
-              <div className="w-1/3 md:flex flex-col">
-                <label
-                  htmlFor="phoneNumber"
-                  className="text-xl lg:text-2xl font-quicksand"
-                >
-                  Phone Number <span className="asterisk">*</span>
-                </label>
-                <input
-                  type=""
-                  name="phoneNumber"
-                  id="phoneNumber"
-                  value={phoneNumber}
-                  onChange={(e) => setPhoneNumber(e.target.value)}
-                  className="border outline-none focus:border-slate-600 py-2 rounded-md"
-                />
-              </div>
-              <div className="w-1/3 md:flex flex-col">
-                <label
-                  htmlFor="postalCode"
-                  className="text-xl lg:text-2xl font-quicksand"
-                >
-                  Postal Code <span className="asterisk">*</span>
-                </label>
-                <input
-                  type="text"
-                  name="postalCode"
-                  id="postalCode"
-                  value={postalCode}
-                  onChange={(e) => setPostalCode(e.target.value)}
-                  className="border outline-none focus:border-slate-600 py-2 rounded-md"
-                />
-              </div>
-            </div>
-          </div>
-
           <h1 className="mt-14 text-xl lg:text-2xl font-quicksand font-semibold my-5">
             Book Shipping Estimator
           </h1>
@@ -1215,8 +728,9 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
                   id="numberOfCopies"
                   className="border outline-none focus:border-slate-500 py-2 ps-2"
                   min="1"
-                  onChange={(e) => setQuantity(+e.target.value)}
-                  disabled={!shippingAddress && !sku}
+                  value={quantity}
+                  onChange={(e) => handleNumberOfCopiesChange(e.target.value)}
+                  disabled={!podPackageId}
                 />
               </div>
               <div className="flex flex-col gap-5">
@@ -1227,11 +741,11 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
                   id="country"
                   name="country"
                   className="border outline-none p-2 w-full"
+                  value={country}
                   onChange={(e) => {
                     getShippingOptions(
                       e.target.value as keyof typeof countryList
                     );
-                    setSelectedShippingOption(null);
                   }}
                   disabled={!quantity}
                 >
@@ -1254,12 +768,14 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
                       ? " border-l-teal border-l-4"
                       : ""
                   }`}
-                  onClick={() => handleEstimatedCost(shippingOption.level)}
+                  onClick={() =>
+                    handleShippingOptionChange(shippingOption.level)
+                  }
                 >
                   <input
                     type="radio"
                     id={`${shippingOption.level}`}
-                    name="coverFinish"
+                    name="shippingOption"
                     value={`${shippingOption.level}`}
                     checked={
                       selectedShippingOption === `${shippingOption.level}`
@@ -1278,7 +794,7 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
                         </p>
                         <p>
                           Delivered in {shippingOption.total_days_min}-
-                          {shippingOption.total_days_min} days
+                          {shippingOption.total_days_max} days
                         </p>
                       </div>
                       <div className="me-4 font-semibold">
@@ -1292,74 +808,200 @@ const OrderBook = ({ params }: { params: OrderBookValues }) => {
             ))}
           </div>
         </div>
-        <div>
-          <h1 className="mt-14 text-xl lg:text-2xl font-quicksand font-semibold my-5">
-            Total Estimated Pricing
-          </h1>
-          <hr className="border-1 border-black"></hr>
 
-          <div className="w-2/4 border ml-auto my-5">
-            <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-              <table className="w-full text-sm text-left rtl:text-right">
-                <tbody>
-                  <tr className="bg-white border-b">
-                    <th
-                      scope="row"
-                      className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap"
-                    >
-                      Book Printing
-                    </th>
-                    <td className="px-6 py-4 font-bold">
-                      {costs?.currency}{" "}
-                      {costs && costs?.line_item_costs[0]?.total_cost_incl_tax}
-                    </td>
-                  </tr>
-                  <tr className="bg-white border-b">
-                    <th
-                      scope="row"
-                      className="px-6 py-4 font-medium whitespace-nowrap"
-                    >
-                      Shipping & Handling
-                    </th>
-                    <td className="px-6 py-4 font-bold">
-                      {costs?.currency}{" "}
-                      {costs &&
-                        +costs?.shipping_cost?.total_cost_incl_tax +
-                          +costs?.fulfillment_cost?.total_cost_incl_tax}
-                    </td>
-                  </tr>
-                </tbody>
-                <tfoot>
-                  <tr className="font-semibold text-gray-900">
-                    <th scope="row" className="px-6 py-3 text-base">
-                      Total
-                    </th>
-                    <td className="px-6 py-3 font-bold">
-                      {costs?.currency}{" "}
-                      {costs &&
-                        Math.round(
-                          (+costs?.line_item_costs[0]?.total_cost_incl_tax +
-                            +costs?.shipping_cost?.total_cost_incl_tax +
-                            +costs?.fulfillment_cost?.total_cost_incl_tax) *
-                            100
-                        ) / 100}
-                    </td>
-                  </tr>
-                </tfoot>
-              </table>
+        {selectedShippingOption && (
+          <div className="flex flex-col gap-5">
+            <h1 className="text-xl lg:text-2xl font-quicksand font-semibold">
+              Shipping Address
+            </h1>
+            <hr className="border-1 border-black"></hr>
+
+            <div className="md:flex flex-col mt-5">
+              <label
+                htmlFor="fullName"
+                className="text-xl lg:text-2xl font-quicksand"
+              >
+                Full Name <span className="asterisk">*</span>
+              </label>
+              <input
+                type="text"
+                name="fullName"
+                id="fullName"
+                ref={fullName}
+                className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+              />
+            </div>
+
+            <div className="md:flex flex-col mt-5">
+              <label
+                htmlFor="streetAddress"
+                className="text-xl lg:text-2xl font-quicksand"
+              >
+                Street Address <span className="asterisk">*</span>
+              </label>
+              <input
+                type="text"
+                name="streetAddress"
+                id="streetAddress"
+                ref={streetAddress}
+                className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+              />
+            </div>
+
+            <div className="flex gap-10">
+              <div className="w-1/2 md:flex flex-col">
+                <label
+                  htmlFor="city"
+                  className="text-xl lg:text-2xl font-quicksand"
+                >
+                  City <span className="asterisk">*</span>
+                </label>
+                <input
+                  type="text"
+                  name="city"
+                  id="city"
+                  ref={city}
+                  className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+                />
+              </div>
+              <div className="w-1/2 md:flex flex-col">
+                <label
+                  htmlFor="state"
+                  className="text-xl lg:text-2xl font-quicksand"
+                >
+                  State <span className="asterisk">*</span>
+                </label>
+                <input
+                  type="text"
+                  name="state"
+                  id="state"
+                  ref={state}
+                  className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+                />
+              </div>
+            </div>
+
+            <div className="flex justify-between gap-5 my-5">
+              <div className="w-1/3 md:flex flex-col">
+                <label
+                  htmlFor="email"
+                  className="text-xl lg:text-2xl font-quicksand"
+                >
+                  Email <span className="asterisk">*</span>
+                </label>
+                <input
+                  type="email"
+                  name="email"
+                  id="email"
+                  ref={email}
+                  className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+                  required
+                />
+              </div>
+              <div className="w-1/3 md:flex flex-col">
+                <label
+                  htmlFor="phoneNumber"
+                  className="text-xl lg:text-2xl font-quicksand"
+                >
+                  Phone Number <span className="asterisk">*</span>
+                </label>
+                <input
+                  type=""
+                  name="phoneNumber"
+                  id="phoneNumber"
+                  ref={phoneNumber}
+                  className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+                />
+              </div>
+              <div className="w-1/3 md:flex flex-col">
+                <label
+                  htmlFor="postalCode"
+                  className="text-xl lg:text-2xl font-quicksand"
+                >
+                  Postal Code <span className="asterisk">*</span>
+                </label>
+                <input
+                  type="text"
+                  name="postalCode"
+                  id="postalCode"
+                  ref={postalCode}
+                  className="ps-3 border outline-none focus:border-slate-600 py-2 rounded-md"
+                />
+              </div>
+            </div>
+            <div className="my-10 text-center">
+              <Button
+                intent="teal"
+                className="text font-bold"
+                onClick={handleEstimatedCost}
+              >
+                Get Quote
+              </Button>
             </div>
           </div>
-        </div>
+        )}
+        {costs && (
+          <div>
+            <div>
+              <h1 className="mt-14 text-xl lg:text-2xl font-quicksand font-semibold my-5">
+                Total Estimated Pricing
+              </h1>
+              <hr className="border-1 border-black"></hr>
 
-        <div className="my-10 text-center">
-          <Button
-            intent="teal"
-            className="text font-bold"
-            onClick={makePayment}
-          >
-            Make Payment
-          </Button>
-        </div>
+              <div className="w-2/4 border ml-auto my-5">
+                <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
+                  <table className="w-full text-sm text-left rtl:text-right">
+                    <tbody>
+                      <tr className="bg-white border-b">
+                        <th
+                          scope="row"
+                          className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap"
+                        >
+                          Book Printing ( {quantity} Copies )
+                        </th>
+                        <td className="px-6 py-4 font-bold">
+                          {costs?.currency} {costs && printingCost}
+                        </td>
+                      </tr>
+                      <tr className="bg-white border-b">
+                        <th
+                          scope="row"
+                          className="px-6 py-4 font-medium whitespace-nowrap"
+                        >
+                          Shipping & Handling
+                        </th>
+                        <td className="px-6 py-4 font-bold">
+                          {costs?.currency} {costs && shippingCost}
+                        </td>
+                      </tr>
+                    </tbody>
+                    <tfoot>
+                      <tr className="font-semibold text-gray-900">
+                        <th scope="row" className="px-6 py-3 text-base">
+                          Total ( {quantity} Copies )
+                        </th>
+                        <td className="px-6 py-3 font-bold">
+                          {costs?.currency} {costs && totalCost}
+                        </td>
+                      </tr>
+                    </tfoot>
+                  </table>
+                </div>
+              </div>
+            </div>
+
+            <div className="my-10 text-center">
+              <Button
+                intent="teal"
+                className="text font-bold"
+                onClick={makePayment}
+                disabled={!totalCost}
+              >
+                Make Payment
+              </Button>
+            </div>
+          </div>
+        )}
       </div>
     </section>
   );
